address typescript comments and update to absolute import paths
