+++ frontend/src/metabase/notifications/channels/SlackChannelField.tsx
@@ -3,36 +3,28 @@ import { useState } from "react";
 import { t } from "ttag";
 
 import CS from "metabase/css/core/index.css";
+import {
+  findChannelId,
+  getDisplayNames,
+} from "metabase/notifications/channels/utils";
 import { getApplicationName } from "metabase/selectors/whitelabel";
 import { Autocomplete } from "metabase/ui";
 import { useSelector } from "metabase/utils/redux";
-import type {
-  Channel,
-  ChannelSpec,
-  SlackChannelOption,
-} from "metabase-types/api";
+import type { Channel, ChannelSpec } from "metabase-types/api";
 
 const CHANNEL_FIELD_NAME = "channel";
 const CHANNEL_PREFIX = "#";
 const USER_PREFIX = "@";
-
-export function getDisplayNames(options: SlackChannelOption[]): string[] {
-  return options.map((o) => o.displayName);
-}
-
-export function findChannelId(
-  options: SlackChannelOption[],
-  displayName: string,
-): string | undefined {
-  return options.find((o) => o.displayName === displayName)?.id;
-}
 
 const ALLOWED_PREFIXES = [CHANNEL_PREFIX, USER_PREFIX];
 
 interface SlackChannelFieldProps {
   channel: Channel;
   channelSpec: ChannelSpec;
-  onChannelPropertyChange: any;
+  onChannelPropertyChange: (
+    key: string,
+    value: Record<string, string | boolean>,
+  ) => void;
 }
 
 export const SlackChannelField = ({
@@ -46,7 +38,7 @@ export const SlackChannelField = ({
   const channelField = channelSpec.fields?.find(
     (field) => field.name === CHANNEL_FIELD_NAME,
   );
-  const slackOptions = (channelField?.options ?? []) as SlackChannelOption[];
+  const slackOptions = channelField?.options ?? [];
   const displayNames = getDisplayNames(slackOptions);
   const value = String(channel?.details?.[CHANNEL_FIELD_NAME] ?? "");
 
@@ -56,7 +48,7 @@ export const SlackChannelField = ({
     onChannelPropertyChange("details", {
       ...restDetails,
       [CHANNEL_FIELD_NAME]: value,
-      ...(newChannelId != null && { channel_id: newChannelId }),
+      ...(newChannelId && { channel_id: newChannelId }),
     });
   };
