+++ frontend/src/metabase/notifications/channels/SlackChannelField.unit.spec.tsx
@@ -2,56 +2,12 @@ import userEvent from "@testing-library/user-event";
 import { useState } from "react";
 
 import { renderWithProviders, screen } from "__support__/ui";
+import { SlackChannelField } from "metabase/notifications/channels/SlackChannelField";
 import type {
   Channel,
   ChannelSpec,
   SlackChannelOption,
 } from "metabase-types/api";
-
-import {
-  SlackChannelField,
-  findChannelId,
-  getDisplayNames,
-} from "./SlackChannelField";
-
-describe("getDisplayNames", () => {
-  it("extracts display names from enriched options", () => {
-    const options: SlackChannelOption[] = [
-      { displayName: "#general", id: "C001" },
-      { displayName: "#random", id: "C002" },
-      { displayName: "@user1", id: "U001" },
-    ];
-    expect(getDisplayNames(options)).toEqual(["#general", "#random", "@user1"]);
-  });
-
-  it("returns empty array for empty input", () => {
-    expect(getDisplayNames([])).toEqual([]);
-  });
-});
-
-describe("findChannelId", () => {
-  const options: SlackChannelOption[] = [
-    { displayName: "#general", id: "C001" },
-    { displayName: "#random", id: "C002" },
-    { displayName: "@user1", id: "U001" },
-  ];
-
-  it("returns the ID for a known channel name", () => {
-    expect(findChannelId(options, "#general")).toBe("C001");
-  });
-
-  it("returns undefined for an unknown channel name", () => {
-    expect(findChannelId(options, "#nonexistent")).toBeUndefined();
-  });
-
-  it("returns undefined for empty string", () => {
-    expect(findChannelId(options, "")).toBeUndefined();
-  });
-
-  it("returns undefined for empty options", () => {
-    expect(findChannelId([], "#general")).toBeUndefined();
-  });
-});
 
 const SLACK_OPTIONS: SlackChannelOption[] = [
   { displayName: "#general", id: "C001" },
@@ -69,21 +25,26 @@ const CHANNEL_SPEC: ChannelSpec = {
     {
       name: "channel",
       displayName: "Post to",
-      options: SLACK_OPTIONS as any,
+      options: SLACK_OPTIONS,
       required: true,
     },
   ],
 };
+
+type OnChannelPropertyChange = (
+  key: string,
+  value: Record<string, string | boolean>,
+) => void;
 
 function StatefulWrapper({
   initialChannel,
   spy,
 }: {
   initialChannel: Channel;
-  spy: jest.Mock;
+  spy: jest.MockedFunction<OnChannelPropertyChange>;
 }) {
   const [channel, setChannel] = useState(initialChannel);
-  const onChannelPropertyChange = (prop: string, value: any) => {
+  const onChannelPropertyChange: OnChannelPropertyChange = (prop, value) => {
     spy(prop, value);
     if (prop === "details") {
       setChannel((prev) => ({ ...prev, details: value }));
@@ -99,7 +60,10 @@ function StatefulWrapper({
 }
 
 function setup({ channel }: { channel: Channel }) {
-  const spy = jest.fn();
+  const spy = jest.fn<
+    ReturnType<OnChannelPropertyChange>,
+    Parameters<OnChannelPropertyChange>
+  >();
   renderWithProviders(<StatefulWrapper initialChannel={channel} spy={spy} />);
   return { spy };
 }
@@ -117,7 +81,6 @@ describe("SlackChannelField stale channel_id prevention", () => {
     await userEvent.clear(input);
     await userEvent.type(input, "#random");
 
-    // The last call should have the NEW channel_id, not the old one
     const lastCall = spy.mock.calls[spy.mock.calls.length - 1];
     const [propName, details] = lastCall;
     expect(propName).toBe("details");
